Consolidate UiComponent2 classes (non-builder/builder-only) into a single class
