+++ lib/over_react.dart
@@ -45,9 +45,11 @@ export 'src/component/dummy_component.dart';
 export 'src/component/prop_mixins.dart';
 export 'src/component/prop_typedefs.dart';
 export 'src/component/resize_sensor.dart';
-export 'src/component_declaration/flux_component.dart';
+export 'src/component_declaration/annotations.dart';
+export 'src/component_declaration/builder_helpers.dart' hide GeneratedErrorMessages;
+export 'src/component_declaration/component_base_2.dart';
 export 'src/component_declaration/built_redux_component.dart';
-export 'src/component_declaration/builder_helpers.dart';
+export 'src/component_declaration/flux_component.dart';
 export 'src/util/character_constants.dart';
 export 'src/util/class_names.dart';
 export 'src/util/constants_base.dart';
